+++ mobile/android/android-components/components/feature/toolbar/src/main/java/mozilla/components/feature/toolbar/ToolbarAutocompleteFeature.kt
@@ -4,42 +4,39 @@
 
 package mozilla.components.feature.toolbar
 
-import mozilla.components.browser.domains.autocomplete.DomainAutocompleteProvider
-import mozilla.components.browser.domains.autocomplete.DomainAutocompleteResult
+import androidx.annotation.VisibleForTesting
+import kotlinx.coroutines.CoroutineScope
 import mozilla.components.concept.engine.Engine
-import mozilla.components.concept.storage.HistoryAutocompleteResult
-import mozilla.components.concept.storage.HistoryStorage
-import mozilla.components.concept.toolbar.AutocompleteResult
+import mozilla.components.concept.toolbar.AutocompleteProvider
 import mozilla.components.concept.toolbar.Toolbar
+import java.util.SortedSet
 
 /**
  * Feature implementation for connecting a toolbar with a list of autocomplete providers.
  *
- * @property toolbar the [Toolbar] to connect to autocomplete providers.
- * @property engine (optional) instance of a browser [Engine] to issue
+ * @param toolbar the [Toolbar] to connect to autocomplete providers.
+ * @param engine (optional) instance of a browser [Engine] to issue
  * [Engine.speculativeConnect] calls on successful URL autocompletion.
- * @property shouldAutocomplete (optional) lambda expression that returns true if
+ * @param shouldAutocomplete (optional) lambda expression that returns true if
  * autocomplete is shown. Otherwise, autocomplete is not shown.
+ * @param scope (optional) [CoroutineScope] in which to query autocompletion providers.
  */
 class ToolbarAutocompleteFeature(
     val toolbar: Toolbar,
     val engine: Engine? = null,
     val shouldAutocomplete: () -> Boolean = { true },
 ) {
-    private val historyProviders: MutableList<HistoryStorage> = mutableListOf()
-    private val domainProviders: MutableList<DomainAutocompleteProvider> = mutableListOf()
+    @VisibleForTesting
+    internal var autocompleteProviders: SortedSet<AutocompleteProvider> = sortedSetOf()
 
     init {
         toolbar.setAutocompleteListener { query, delegate ->
-            if (!shouldAutocomplete()) {
+            if (!shouldAutocomplete() || autocompleteProviders.isEmpty() || query.isBlank()) {
                 delegate.noAutocompleteResult(query)
             } else {
-                val historyResults = historyProviders.asSequence()
-                    .mapNotNull { provider -> provider.getAutocompleteSuggestion(query)?.into() }
-                val domainResults = domainProviders.asSequence()
-                    .mapNotNull { provider -> provider.getAutocompleteSuggestion(query)?.into() }
+                val result = autocompleteProviders
+                    .firstNotNullOfOrNull { it.getAutocompleteSuggestion(query) }
 
-                val result = (historyResults + domainResults).firstOrNull()
                 if (result != null) {
                     delegate.applyAutocompleteResult(result) {
                         engine?.speculativeConnect(result.url)
@@ -51,26 +48,36 @@ class ToolbarAutocompleteFeature(
         }
     }
 
-    fun addHistoryStorageProvider(provider: HistoryStorage) {
-        historyProviders.add(provider)
+    /**
+     * Update the list of providers used for autocompletion results.
+     * Changes will take effect the next time user changes their input.
+     *
+     * @param providers New list of autocomplete providers.
+     * The list can be empty in which case autocompletion will be disabled until there is at least one provider.
+     */
+    @Synchronized
+    fun updateAutocompleteProviders(providers: List<AutocompleteProvider>) {
+        autocompleteProviders.clear()
+        autocompleteProviders.addAll(providers)
     }
 
-    fun addDomainProvider(provider: DomainAutocompleteProvider) {
-        domainProviders.add(provider)
+    /**
+     * Adds the specified provider to the current list of providers.
+     *
+     * @return `true` if the provider has been added, `false` if the provider already exists.
+     */
+    @Synchronized
+    fun addAutocompleteProvider(provider: AutocompleteProvider): Boolean {
+        return autocompleteProviders.add(provider)
     }
 
-    private fun HistoryAutocompleteResult.into() = AutocompleteResult(
-        input = input,
-        text = text,
-        url = url,
-        source = source,
-        totalItems = totalItems,
-    )
-    private fun DomainAutocompleteResult.into() = AutocompleteResult(
-        input = input,
-        text = text,
-        url = url,
-        source = source,
-        totalItems = totalItems,
-    )
+    /**
+     * Remove an autocomplete provider from the current providers list.
+     *
+     * @return `true` if the provider has been removed, `false` if the provider could not be found.
+     */
+    @Synchronized
+    fun removeAutocompleteProvider(provider: AutocompleteProvider): Boolean {
+        return autocompleteProviders.remove(provider)
+    }
 }