+++ mobile/android/android-components/components/feature/toolbar/src/test/java/mozilla/components/feature/toolbar/ToolbarAutocompleteFeatureTest.kt
@@ -10,28 +10,30 @@ import mozilla.components.browser.domains.Domain
 import mozilla.components.browser.domains.autocomplete.BaseDomainAutocompleteProvider
 import mozilla.components.browser.domains.autocomplete.DomainList
 import mozilla.components.concept.engine.Engine
-import mozilla.components.concept.storage.HistoryAutocompleteResult
-import mozilla.components.concept.storage.HistoryStorage
 import mozilla.components.concept.toolbar.AutocompleteDelegate
+import mozilla.components.concept.toolbar.AutocompleteProvider
 import mozilla.components.concept.toolbar.AutocompleteResult
 import mozilla.components.concept.toolbar.Toolbar
 import mozilla.components.support.test.any
 import mozilla.components.support.test.argumentCaptor
 import mozilla.components.support.test.eq
 import mozilla.components.support.test.mock
+import org.junit.Assert.assertEquals
+import org.junit.Assert.assertFalse
 import org.junit.Assert.assertNotNull
+import org.junit.Assert.assertTrue
 import org.junit.Assert.fail
 import org.junit.Test
 import org.junit.runner.RunWith
 import org.mockito.Mockito.doReturn
 import org.mockito.Mockito.never
 import org.mockito.Mockito.reset
+import org.mockito.Mockito.spy
 import org.mockito.Mockito.times
 import org.mockito.Mockito.verify
 
 @RunWith(AndroidJUnit4::class)
 class ToolbarAutocompleteFeatureTest {
-
     class TestToolbar : Toolbar {
         override var highlight: Toolbar.Highlight = Toolbar.Highlight.NONE
         override var siteTrackingProtection: Toolbar.SiteTrackingProtection =
@@ -157,25 +159,25 @@ class ToolbarAutocompleteFeatureTest {
     }
 
     @Test
-    fun `feature can be configured with providers`() {
+    fun `feature can be configured with providers`() = runTest {
         val toolbar = TestToolbar()
         var feature = ToolbarAutocompleteFeature(toolbar)
         val autocompleteDelegate: AutocompleteDelegate = mock()
 
-        var history: HistoryStorage = mock()
-        val domains = object : BaseDomainAutocompleteProvider(DomainList.CUSTOM, { emptyList() }) {
+        var history: AutocompleteProvider = mock()
+        val domains = object : BaseDomainAutocompleteProvider(DomainList.CUSTOM, { emptyList() }, 22) {
             fun testDomains(list: List<Domain>) {
                 domains = list
             }
         }
 
         // Can autocomplete with just an empty history provider.
-        feature.addHistoryStorageProvider(history)
+        feature.addAutocompleteProvider(history)
         verifyNoAutocompleteResult(toolbar, autocompleteDelegate, "hi")
 
         // Can autocomplete with a non-empty history provider.
         doReturn(
-            HistoryAutocompleteResult(
+            AutocompleteResult(
                 input = "mo",
                 text = "mozilla.org",
                 url = "https://www.mozilla.org",
@@ -200,7 +202,7 @@ class ToolbarAutocompleteFeatureTest {
 
         // Can autocomplete with just an empty domain provider.
         feature = ToolbarAutocompleteFeature(toolbar)
-        feature.addDomainProvider(domains)
+        feature.addAutocompleteProvider(domains)
 
         verifyNoAutocompleteResult(toolbar, autocompleteDelegate, "hi")
 
@@ -226,9 +228,9 @@ class ToolbarAutocompleteFeatureTest {
         )
 
         // Can autocomplete with empty history and domain providers.
-        history = mock()
+        history = spy(AutocompleteProviderFake()) // use a real object so that the comparator will work
         domains.testDomains(listOf())
-        feature.addHistoryStorageProvider(history)
+        feature.addAutocompleteProvider(history)
 
         verifyNoAutocompleteResult(toolbar, autocompleteDelegate, "hi")
 
@@ -255,7 +257,7 @@ class ToolbarAutocompleteFeatureTest {
         )
 
         doReturn(
-            HistoryAutocompleteResult(
+            AutocompleteResult(
                 input = "mo",
                 text = "mozilla.org",
                 url = "https://www.mozilla.org",
@@ -295,7 +297,7 @@ class ToolbarAutocompleteFeatureTest {
     fun `feature triggers speculative connect for results if engine provided`() = runTest {
         val toolbar = TestToolbar()
         val engine: Engine = mock()
-        var feature = ToolbarAutocompleteFeature(toolbar, engine)
+        val feature = ToolbarAutocompleteFeature(toolbar, engine) { true }
         val autocompleteDelegate: AutocompleteDelegate = mock()
 
         val domains = object : BaseDomainAutocompleteProvider(DomainList.CUSTOM, { emptyList() }) {
@@ -304,7 +306,7 @@ class ToolbarAutocompleteFeatureTest {
             }
         }
         domains.testDomains(listOf(Domain.create("https://www.mozilla.org")))
-        feature.addDomainProvider(domains)
+        feature.addAutocompleteProvider(domains)
 
         toolbar.autocompleteFilter!!.invoke("mo", autocompleteDelegate)
 
@@ -329,7 +331,7 @@ class ToolbarAutocompleteFeatureTest {
             }
         }
         domains.testDomains(listOf(Domain.create("https://www.mozilla.org")))
-        feature.addDomainProvider(domains)
+        feature.addAutocompleteProvider(domains)
 
         toolbar.autocompleteFilter!!.invoke("mo", autocompleteDelegate)
 
@@ -344,6 +346,169 @@ class ToolbarAutocompleteFeatureTest {
         verify(engine, never()).speculativeConnect("https://www.mozilla.org")
         callbackCaptor.value.invoke()
         verify(engine).speculativeConnect("https://www.mozilla.org")
+    }
+
+    @Test
+    fun `GIVEN no autocomplete providers WHEN checking for autocomplete results THEN silently fail with no results`() = runTest {
+        val toolbar = TestToolbar()
+        val engine: Engine = mock()
+        val feature = ToolbarAutocompleteFeature(toolbar, engine) { true }
+        val autocompleteDelegate: AutocompleteDelegate = mock()
+
+        val domains = object : BaseDomainAutocompleteProvider(DomainList.CUSTOM, { emptyList() }) {
+            fun testDomains(list: List<Domain>) {
+                domains = list
+            }
+        }
+        domains.testDomains(listOf(Domain.create("https://www.mozilla.org")))
+
+        feature.addAutocompleteProvider(domains)
+        toolbar.autocompleteFilter!!.invoke("mo", autocompleteDelegate)
+        val callbackCaptor = argumentCaptor<() -> Unit>()
+        verify(autocompleteDelegate, times(1)).applyAutocompleteResult(any(), callbackCaptor.capture())
+        verify(engine, never()).speculativeConnect("https://www.mozilla.org")
+        callbackCaptor.value.invoke()
+        verify(engine).speculativeConnect("https://www.mozilla.org")
+
+        // After checking the results for when a provider exists test what happens when no providers exist.
+        feature.removeAutocompleteProvider(domains)
+        toolbar.autocompleteFilter!!.invoke("mo", autocompleteDelegate)
+        verify(autocompleteDelegate, times(1)).noAutocompleteResult(any())
+        verify(engine, times(1)).speculativeConnect("https://www.mozilla.org")
+    }
+
+    @Test
+    fun `GIVEN no initial autocomplete providers and one is added WHEN checking for autocomplete results THEN return autocomplete suggestions`() = runTest {
+        val toolbar = TestToolbar()
+        val engine: Engine = mock()
+        val feature = ToolbarAutocompleteFeature(toolbar, engine) { true }
+        val autocompleteDelegate: AutocompleteDelegate = mock()
+
+        val domains = object : BaseDomainAutocompleteProvider(DomainList.CUSTOM, { emptyList() }) {
+            fun testDomains(list: List<Domain>) {
+                domains = list
+            }
+        }
+        domains.testDomains(listOf(Domain.create("https://www.mozilla.org")))
+
+        toolbar.autocompleteFilter!!.invoke("mo", autocompleteDelegate)
+        verify(autocompleteDelegate, times(1)).noAutocompleteResult(any())
+        verify(engine, never()).speculativeConnect("https://www.mozilla.org")
+
+        // After checking no results for when no providers exist test what happens when a new one is added.
+        feature.addAutocompleteProvider(domains)
+        toolbar.autocompleteFilter!!.invoke("mo", autocompleteDelegate)
+        val callbackCaptor = argumentCaptor<() -> Unit>()
+        verify(autocompleteDelegate, times(1)).applyAutocompleteResult(any(), callbackCaptor.capture())
+        verify(engine, never()).speculativeConnect("https://www.mozilla.org")
+        callbackCaptor.value.invoke()
+        verify(engine).speculativeConnect("https://www.mozilla.org")
+    }
+
+    @Test
+    fun `GIVEN providers exist WHEN a new one is added THEN they are sorted by their priority`() {
+        val feature = ToolbarAutocompleteFeature(mock())
+
+        val provider1 = AutocompleteProviderFake(autocompletePriority = 11).also {
+            assertTrue(feature.addAutocompleteProvider(it))
+        }
+        val provider2 = AutocompleteProviderFake(autocompletePriority = 22).also {
+            assertTrue(feature.addAutocompleteProvider(it))
+        }
+        val provider3 = AutocompleteProviderFake(autocompletePriority = 3).also {
+            assertTrue(feature.addAutocompleteProvider(it))
+        }
+
+        assertEquals(
+            listOf(provider3, provider1, provider2),
+            feature.autocompleteProviders.toList(),
+        )
+    }
+
+    @Test
+    fun `GIVEN providers exist WHEN trying to add an existing one THEN avoid adding`() {
+        val feature = ToolbarAutocompleteFeature(mock())
+        val provider1 = AutocompleteProviderFake(autocompletePriority = 11).also {
+            feature.addAutocompleteProvider(it)
+        }
+        val provider2 = AutocompleteProviderFake(autocompletePriority = 22).also {
+            feature.addAutocompleteProvider(it)
+        }
+        val provider3 = AutocompleteProviderFake(autocompletePriority = 3).also {
+            feature.addAutocompleteProvider(it)
+        }
+        val provider4 = AutocompleteProviderFake(autocompletePriority = 15).also {
+            feature.addAutocompleteProvider(it)
+        }
+
+        assertTrue(feature.removeAutocompleteProvider(provider4))
+
+        assertEquals(
+            listOf(provider3, provider1, provider2),
+            feature.autocompleteProviders.toList(),
+        )
+    }
+
+    @Test
+    fun `GIVEN providers don't exist WHEN trying to remove one THEN avoid fail gracefully`() {
+        val feature = ToolbarAutocompleteFeature(mock())
+        val provider1 = AutocompleteProviderFake(autocompletePriority = 11).also {
+            feature.addAutocompleteProvider(it)
+        }
+        val provider2 = AutocompleteProviderFake(autocompletePriority = 22).also {
+            feature.addAutocompleteProvider(it)
+        }
+        val provider3 = AutocompleteProviderFake(autocompletePriority = 3)
+
+        assertFalse(feature.removeAutocompleteProvider(provider3))
+
+        assertEquals(
+            listOf(provider1, provider2),
+            feature.autocompleteProviders.toList(),
+        )
+    }
+
+    @Test
+    fun `GIVEN providers exist WHEN removing one THEN keep the other sorted`() {
+        val feature = ToolbarAutocompleteFeature(mock())
+        val provider1 = AutocompleteProviderFake(autocompletePriority = 11).also {
+            assertTrue(feature.addAutocompleteProvider(it))
+        }
+        val provider2 = AutocompleteProviderFake(autocompletePriority = 22).also {
+            assertTrue(feature.addAutocompleteProvider(it))
+        }
+
+        assertFalse(feature.addAutocompleteProvider(provider1))
+
+        assertEquals(
+            listOf(provider1, provider2),
+            feature.autocompleteProviders.toList(),
+        )
+    }
+
+    @Test
+    fun `GIVEN providers exist WHEN when they are updated THEN the old ones are replaced by new ones sorted by priority`() {
+        val feature = ToolbarAutocompleteFeature(mock())
+        feature.autocompleteProviders = sortedSetOf(
+            AutocompleteProviderFake(autocompletePriority = 11),
+            AutocompleteProviderFake(autocompletePriority = 22),
+        )
+        val provider1 = AutocompleteProviderFake(autocompletePriority = 11).also {
+            feature.addAutocompleteProvider(it)
+        }
+        val provider2 = AutocompleteProviderFake(autocompletePriority = 2).also {
+            feature.addAutocompleteProvider(it)
+        }
+        val provider3 = AutocompleteProviderFake().also {
+            feature.addAutocompleteProvider(it)
+        }
+
+        feature.updateAutocompleteProviders(listOf(provider1, provider2, provider3))
+
+        assertEquals(
+            listOf(provider3, provider2, provider1),
+            feature.autocompleteProviders.toList(),
+        )
     }
 
     @Suppress("SameParameterValue")
@@ -362,4 +527,17 @@ class ToolbarAutocompleteFeatureTest {
         verify(autocompleteDelegate, never()).noAutocompleteResult(query)
         reset(autocompleteDelegate)
     }
+}
+
+/**
+ * Empty implementation of [AutocompleteProvider].
+ * [getAutocompleteSuggestion] will return `null` by default.
+ *
+ * @param resultToReturn Optional nullable [AutocompleteResult] to return for all queries.
+ */
+private class AutocompleteProviderFake(
+    val resultToReturn: AutocompleteResult? = null,
+    override val autocompletePriority: Int = 0,
+) : AutocompleteProvider {
+    override suspend fun getAutocompleteSuggestion(query: String) = resultToReturn
 }